Create unified StellarRequestReviewDialog which combines handling payment and transaction requests
